Use `usize` as the Notify task identifier.
Using u64 isn't a clear win and it will usually represent a pointer
anyway.

+++ benches/poll.rs
@@ -13,7 +13,7 @@ fn notify_noop() -> NotifyHandle {
     struct Noop;
 
     impl Notify for Noop {
-        fn notify(&self, _id: u64) {}
+        fn notify(&self, _id: usize) {}
     }
 
     const NOOP : &'static Noop = &Noop;
